Created editor
Create EditTextFragment  and its modelView with Factory.
When SetOnClickListener each items was implemented via send lambda to TaskItemAdapter and getting value via Observer.
Created fragment_edit_task xml and navigation between fragment_task and fragment_edit_text.

+++ app/src/main/java/production/zhandos/tasks/TaskDao.kt
@@ -15,7 +15,7 @@ interface TaskDao {
     suspend fun delete(task: Task)
 
     @Query("SELECT * FROM task_table WHERE taskId=:taskId")
-    fun getById(taskId: Int): LiveData<Task>
+    fun getById(taskId: Long): LiveData<Task>
 
     @Query("SELECT * FROM task_table ORDER BY taskId DESC")
     fun getAll(): LiveData<List<Task>>
